# docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Violeta API",
        "description": "Documentación de endpoints del backend Violeta\n\n**Autenticación - Dos métodos disponibles:**\n\n**1. Cookie/Session (Para aplicaciones web):**\n- Ejecuta POST /api/v1/auth/login con tus credenciales\n- La sesión se crea automáticamente (cookie JSESSIONID)\n- Los siguientes requests usan la sesión automáticamente\n- No necesitas configurar nada en \"Authorize\"\n\n**2. JWT Token (Para APIs, móviles, scripts):**\n- Ejecuta POST /api/v1/auth/login/jwt con tus credenciales\n- Recibes un token JWT en el campo \"token\" del response\n- Haz clic en \"Authorize\" arriba a la derecha\n- Pega el token en \"bearerAuth\" (sin agregar \"Bearer \")\n- El token expira en 4 horas\n\n**¿Cuál usar?**\n- Aplicación web React → Cookie/Session (endpoint /login)\n- App móvil/API/Scripts → JWT (endpoint /login/jwt)\n",
        "contact": {
            "name": "Equipo Desarrollo",
            "email": "dev@lavioleta.com"
        },
        "license": {
            "name": "Propietario",
            "url": "https://lavioleta.com"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:6820",
            "description": "Desarrollo Local"
        }
    ],
    "security": [
        {
            "cookieAuth": []
        },
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/api/v1/security/privilegios/{objeto}": {
            "get": {
                "tags": [
                    "authorization-controller"
                ],
                "operationId": "listarPorObjeto",
                "parameters": [
                    {
                        "name": "objeto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PrivilegioDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/security/privilegios/{objeto}/{privilegio}": {
            "get": {
                "tags": [
                    "authorization-controller"
                ],
                "operationId": "revisar",
                "parameters": [
                    {
                        "name": "objeto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "privilegio",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PrivilegeCheckResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/security/privilegios/cache/{usuario}": {
            "delete": {
                "tags": [
                    "authorization-controller"
                ],
                "operationId": "flushCacheUsuario",
                "parameters": [
                    {
                        "name": "usuario",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PrivilegioDTO": {
                "type": "object",
                "properties": {
                    "objeto": {
                        "type": "string"
                    },
                    "privilegio": {
                        "type": "string"
                    },
                    "descripcion": {
                        "type": "string"
                    }
                }
            },
            "PrivilegeCheckResponse": {
                "type": "object",
                "properties": {
                    "objeto": {
                        "type": "string"
                    },
                    "privilegio": {
                        "type": "string"
                    },
                    "allowed": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "description": "**Autenticación basada en sesión HTTP (para aplicaciones web)**\n\n1. Ejecuta POST /api/v1/auth/login con tus credenciales\n2. La sesión se crea automáticamente (cookie JSESSIONID)\n3. Las cookies se envían automáticamente en los siguientes requests\n\nIdeal para: Aplicación web React, navegadores\n",
                "name": "JSESSIONID",
                "in": "cookie"
            },
            "bearerAuth": {
                "type": "http",
                "description": "**Autenticación basada en JWT (para APIs, móviles, scripts)**\n\n1. Ejecuta POST /api/v1/auth/login/jwt con tus credenciales\n2. Copia el token del campo \"token\" en el response\n3. Haz clic en \"Authorize\" arriba a la derecha\n4. Pega el token en este campo (sin agregar \"Bearer \")\n5. Usa el token en tus requests: Authorization: Bearer {token}\n\nEl token expira en 4 horas.\n\nIdeal para: Apps móviles, APIs externas, scripts de automatización\n",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}